Updates to UI mockup.
I've updated the UI to be closer to what we discussed this week:

   - The main screen now contains the following:

      - An inset which displays current conditions. The air quality
        data is randomly generated by the
        jimjams.airmonitor.sensordata.SensorDataGenerator class.

      - A button linking to the EMA screen

      - A button to refresh current air quality data

   - The EMA screen contains:

      - The questionnaire based on the EMA outline we got from
        HabitatMap

      - A save button (not yet functional)

      - A cancel button (not yet functional)

The layout of the EMA page still needs some work. I think I need a
Fragment or something to contain the questionnaire, because the save
and cancel buttons are overlapping it.

One other note: I've changed the package structure so that it reflects
the team name instead of an individual member's name.