Updates to existing condition list.
It is now possible to add existing conditions to and remove them from the profile.

Tha app should be able to generate an EcologicalMomentaryAssessment and a Snapshot, but there is not yet any output based on these, so this is untested.

Also got output messages to log working, so that will make debuggery a lot easier.